# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para sistema de gestão de serviços
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  address?: string;
  notes?: string;
}

export interface ClientUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  notes?: string;
}

export interface Service {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  unitPrice: number;
  unit: string;
  createdAt: string;
}

export interface ServiceInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  unitPrice: number;
  /** @minLength 1 */
  unit: string;
}

export interface ServiceUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  unitPrice?: number;
  /** @minLength 1 */
  unit?: string;
}

export interface Material {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  unitPrice: number;
  unit: string;
  stockQuantity: number;
  createdAt: string;
}

export interface MaterialInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  unitPrice: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  stockQuantity?: number;
}

export interface MaterialUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  unitPrice?: number;
  /** @minLength 1 */
  unit?: string;
  /** @minimum 0 */
  stockQuantity?: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  clientId: number;
  clientName: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: OrderStatus;
  totalPrice: number;
  createdAt: string;
  updatedAt?: string;
}

export type OrderDetailStatus = typeof OrderDetailStatus[keyof typeof OrderDetailStatus];


export const OrderDetailStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderServiceItem {
  id: number;
  orderId: number;
  serviceId: number;
  serviceName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export interface OrderMaterialItem {
  id: number;
  orderId: number;
  materialId: number;
  materialName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export interface OrderDetail {
  id: number;
  clientId: number;
  clientName: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: OrderDetailStatus;
  totalPrice: number;
  createdAt: string;
  updatedAt?: string;
  serviceItems: OrderServiceItem[];
  materialItems: OrderMaterialItem[];
}

export type OrderInputStatus = typeof OrderInputStatus[keyof typeof OrderInputStatus];


export const OrderInputStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderInput {
  clientId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: OrderInputStatus;
}

export type OrderUpdateStatus = typeof OrderUpdateStatus[keyof typeof OrderUpdateStatus];


export const OrderUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: OrderUpdateStatus;
}

export interface OrderServiceItemInput {
  serviceId: number;
  /** @minimum 0.01 */
  quantity: number;
  /** @minimum 0 */
  unitPrice?: number;
}

export interface OrderMaterialItemInput {
  materialId: number;
  /** @minimum 0.01 */
  quantity: number;
  /** @minimum 0 */
  unitPrice?: number;
}

export interface Summary {
  totalClients: number;
  totalServices: number;
  totalMaterials: number;
  totalOrders: number;
  pendingOrders: number;
  inProgressOrders: number;
  completedOrders: number;
  cancelledOrders: number;
  totalRevenue: number;
}