# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API para sistema de gestão de serviços
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Services (catalog)
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List all materials
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get a material by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Orders
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Client
    Client:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string

    # Service
...
    Material:
      type: object
      required: [id, name, unitPrice, unit, stockQuantity, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        unit:
          type: string
        stockQuantity:
          type: number
        createdAt:
          type: string
          format: date-time
    MaterialInput:
      type: object
      required: [name, unitPrice, unit]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        unitPrice:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
        stockQuantity:
          type: number
          minimum: 0
    MaterialUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        unitPrice:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
        stockQuantity:
          type: number
          minimum: 0

    # Order
...