# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para sistema de gestão de serviços
 * OpenAPI spec version: 0.1.0
 */

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  address?: string;
  notes?: string;
}