# lib/api-zod/src/generated/types/materialInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para sistema de gestão de serviços
 * OpenAPI spec version: 0.1.0
 */

export interface MaterialInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  unitPrice: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  stockQuantity?: number;
}